dx: split functions in separate files
