+++ .eslintrc.js
@@ -1,10 +1,6 @@
 module.exports = {
   parser: '@typescript-eslint/parser',
-  extends: [
-    'eslint:recommended',
-    'google',
-    'prettier',
-  ],
+  extends: ['eslint:recommended', 'google', 'prettier'],
   env: {
     commonjs: true,
     es6: true,
@@ -25,10 +21,8 @@ module.exports = {
   },
   rules: {
     indent: ['error', 2, { SwitchCase: 1 }],
-    'no-console':
-      process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    'no-debugger':
-      process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'object-curly-spacing': 'off',
     'valid-jsdoc': 'off',
     'require-jsdoc': 'off',
