+++ .github/workflows/npm_publish.yml
@@ -20,7 +20,7 @@ jobs:
         uses: supercharge/redis-github-action@1.1.0
         with:
           redis-version: ${{ matrix.redis-version }}
-      - run: npm install  
+      - run: npm install
       - run: npm test
 
   publish-npm:
