+++ lib/cache/index.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-const mapValues = require("lodash/mapValues");
+const mapValues = require('lodash/mapValues');
 
 const rejectAfterTimeout = (ms) =>
   new Promise((_, reject) => setTimeout(reject, ms));
@@ -46,24 +46,25 @@ module.exports = (strapi, options, logger) => {
         return await store.reset();
       },
     },
-    (fn) => async (...args) => {
-      try {
-        if (!store.ready) {
-          logger.warn("the store is not ready");
+    (fn) =>
+      async (...args) => {
+        try {
+          if (!store.ready) {
+            logger.warn('the store is not ready');
+            return null;
+          }
+          return await Promise.race([
+            fn(...args),
+            rejectAfterTimeout(options.cacheTimeout),
+          ]);
+        } catch (e) {
+          if (e) {
+            logger.warn('there was an error with the store', e);
+          }
+          logger.warn('the cache request timed out');
+
           return null;
         }
-        return await Promise.race([
-          fn(...args),
-          rejectAfterTimeout(options.cacheTimeout),
-        ]);
-      } catch (e) {
-        if (e) {
-          logger.warn("there was an error with the store", e);
-        } 
-        logger.warn("the cache request timed out");
-        
-        return null;
       }
-    }
   );
 };
