+++ lib/cache/stores/mem.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-const LRU = require("lru-cache");
+const LRU = require('lru-cache');
 
 /**
  *
