+++ lib/cache/stores/redis.js
@@ -1,7 +1,7 @@
 // @ts-check
 
-const Redis = require("ioredis");
-const lru = require("redis-lru");
+const Redis = require('ioredis');
+const lru = require('redis-lru');
 
 /**
  *
@@ -11,25 +11,25 @@ const lru = require("redis-lru");
  */
 module.exports = (strapi, options, logger) => {
   const prefix = (key) => `strapi-middleware-cache:${key}`;
-  const unprefix = (key) => key.replace(/^strapi-middleware-cache:/, "");
+  const unprefix = (key) => key.replace(/^strapi-middleware-cache:/, '');
 
   const client = new Redis(options.redisConfig);
   const cache = lru(client, options.max);
   let reconnected = false;
 
-  client.on("ready", () => {
-    logger.debug("redis connection established");
+  client.on('ready', () => {
+    logger.debug('redis connection established');
 
     if (reconnected) {
       cache.reset();
     }
   });
 
-  client.on("error", (error) => {
+  client.on('error', (error) => {
     logger.error(`redis connection failed`, error);
   });
 
-  client.on("reconnecting", (e) => {
+  client.on('reconnecting', (e) => {
     reconnected = true;
   });
 
@@ -54,7 +54,7 @@ module.exports = (strapi, options, logger) => {
       return cache.reset();
     },
     get ready() {
-      return client.status === "ready";
+      return client.status === 'ready';
     },
   };
 };
