+++ lib/utils/arrayEquals.js
@@ -1,0 +1,14 @@
+/**
+ * @param {string[]} a
+ * @param {string[]} b
+ */
+function arrayEquals(a, b) {
+  return (
+    Array.isArray(a) &&
+    Array.isArray(b) &&
+    a.length === b.length &&
+    a.every((val, index) => val === b[index])
+  );
+}
+
+module.exports = arrayEquals;
