+++ lib/utils/index.js
@@ -1,0 +1,17 @@
+module.exports = {
+  arrayEquals: require('./arrayEquals'),
+  createClearCache: require('./createClearCache'),
+  createGetCacheConfig: require('./createGetCacheConfig'),
+  createGetCacheConfigByUid: require('./createGetCacheConfigByUid'),
+  createLogger: require('./createLogger'),
+  createPurge: require('./createPurge'),
+  createPurgeAdmin: require('./createPurgeAdmin'),
+  createRecv: require('./createRecv'),
+  createRouter: require('./createRouter'),
+  generateCacheKey: require('./generateCacheKey'),
+  getCacheConfRegExp: require('./getCacheConfRegExp'),
+  getRelatedModelsUid: require('./getRelatedModelsUid'),
+  getRouteRegExp: require('./getRouteRegExp'),
+  resolveUserOptions: require('./resolveUserOptions'),
+  routeExists: require('./routeExists'),
+};
