+++ tests/stubs/strapi.js
@@ -2,9 +2,9 @@ const Koa = require('koa');
 
 const requests = [];
 
-const reset = () => requests.length = 0;
+const reset = () => (requests.length = 0);
 
-let UID = 1
+let UID = 1;
 
 const strapi = {
   app: new Koa(),
@@ -12,7 +12,7 @@ const strapi = {
     error() {},
     warn() {},
     info() {},
-    debug() {}
+    debug() {},
   },
   models: {
     academy: {
@@ -22,20 +22,17 @@ const strapi = {
         researchers: {
           via: 'academies',
           collection: 'researcher',
-        }
-      }
+        },
+      },
     },
     homepage: {
       kind: 'singleType',
       allAttributes: {
         body: {
           type: 'dynamiczone',
-          components: [
-            'layouts.researchers',
-            'layouts.images',
-          ]
-        }
-      }
+          components: ['layouts.researchers', 'layouts.images'],
+        },
+      },
     },
     researcher: {
       kind: 'collectionType',
@@ -44,55 +41,53 @@ const strapi = {
         academy: {
           via: 'researcher',
           collection: 'academy',
-        }
-      }
-    }
+        },
+      },
+    },
   },
   plugins: {
     'users-permissions': {
       models: {
-        user: {}
-      }
-    }
+        user: {},
+      },
+    },
   },
   components: {
     'fields.researcher': {
       allAttributes: {
         researcher: {
-          model: 'researcher'
-        }
-      }
+          model: 'researcher',
+        },
+      },
     },
     'layouts.researchers': {
       allAttributes: {
         text: {
-          type: 'text'
+          type: 'text',
         },
         researchers: {
           type: 'component',
           repeatable: true,
-          component: 'fields.researcher'
-        }
-      }
-    }
+          component: 'fields.researcher',
+        },
+      },
+    },
   },
   start() {
-    this.app.use(ctx => {
+    this.app.use((ctx) => {
       const data = {
         uid: UID++,
         method: ctx.method,
         url: ctx.url,
-        query: ctx.query
+        query: ctx.query,
       };
-
 
       requests.push(data);
 
       ctx.body = data;
       ctx.status = 200;
     });
-  }
+  },
 };
-
 
 module.exports = { strapi, requests, reset };
